requirements.txt
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
google-genai>=1.0.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
requests>=2.31.0
python-dotenv>=1.0.0
defusedxml>=0.7.0
pyyaml>=6.0
